web-app/requirements.txt
Django>3.0
psycopg2==2.9.9
django-bootstrap5
django-crispy-forms
django-bootstrap-v5
django-bootstrap-select
django-jquery
sendgrid==6.8.0
google-auth==2.27.0
google_auth_oauthlib==1.2.0
google-api-python-client==2.116.0
testresources==2.0.1
launchpadlib==1.10.13
flask==3.0.3